requirements.txt
fastapi
uvicorn
pandas
python-dotenv
google-generativeai==0.3.2
